Made slider and chart reactive

+++ src/apps/blazor-app/Models/SystemInfoModel.cs
@@ -3,7 +3,7 @@ public class SystemInfo
   public string? Size { get; set; }
   public int Usage { get; set; }
   public double Cost { get; set; }
-  public int Coverage { get; set; }
+  public double Coverage { get; set; }
   public int Extra { get; set; }
   public int ExtraSavings { get; set; }
   public int NewBill { get; set; }
